% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_print.R
\name{better_print}
\alias{better_print}
\title{Better Print}
\usage{
better_print(...)
}
\arguments{
\item{...}{Arguments to print}
}
\value{
No return, just prints the output.
}
\description{
This function prints the given arguments in the best possible way. Unlike R's base function,
function can take unlimited arguments, and simply puts them together.
}
